move csv and rst generating functions to exporters file
