+++ threedi_modelchecker/scripts.py
@@ -1,6 +1,3 @@
-import csv
-from io import StringIO
-
 import click
 from threedi_schema import ThreediDatabase
 from threedi_schema.domain.models import DECLARED_MODELS
@@ -69,59 +66,12 @@ def check(sqlite, file, level):
 )
 def export_checks(file, format):
     """Export formatted checks summary to insert in documentation or use elsewhere"""
-
-    def generate_rst_table(checks) -> str:
-        "Generate an RST table to copy into the Sphinx docs with a list of checks"
-        rst_table_string = ""
-        header = (
-            ".. list-table:: Executed checks\n"
-            + "   :widths: 10 20 40\n"
-            + "   :header-rows: 1\n\n"
-            + "   * - Check number\n"
-            + "     - Check level\n"
-            + "     - Check message"
-        )
-        rst_table_string += header
-        for check in checks:
-            # pad error code with leading zeroes so it is always 4 numbers
-            formatted_error_code = str(check.error_code).zfill(4)
-            check_row = (
-                "\n"
-                + f"   * - {formatted_error_code}\n"
-                + f"     - {check.level.name.capitalize()}\n"
-                + f"     - {check.description()}"
-            )
-            rst_table_string += check_row
-        return rst_table_string
-
-    def generate_csv_table(checks) -> str:
-        "Generate an CSV table with a list of checks for use elsewhere"
-        # a StringIO buffer is used so that the CSV can be printed to terminal as well as written to file
-        output_buffer = StringIO()
-        fieldnames = ["error_code", "level", "description"]
-        writer = csv.DictWriter(
-            output_buffer, fieldnames=fieldnames, quoting=csv.QUOTE_NONNUMERIC
-        )
-
-        writer.writeheader()
-
-        for check in checks:
-            writer.writerow(
-                {
-                    "error_code": check.error_code,
-                    "level": check.level.name,
-                    "description": check.description(),
-                }
-            )
-
-        return output_buffer.getvalue()
-
     checks = Config(models=DECLARED_MODELS).checks
 
     if format.lower() == "rst":
-        table = generate_rst_table(checks=checks)
+        table = exporters.generate_rst_table(checks=checks)
     elif format.lower() == "csv":
-        table = generate_csv_table(checks=checks)
+        table = exporters.generate_csv_table(checks=checks)
     if file:
         with open(file, "w") as f:
             f.write(table)
